requirements.txt
imgurpython==1.1.7
Pillow==3.2.0
requests==2.9.1
six==1.10.0
slackclient==1.0.0
websocket-client==0.37.0
pycron==0.40
